+++ packages/list/src/lib/ListSubheader.stories.tsx
@@ -1,8 +1,9 @@
 // Each exported component in the package should have its own stories file
 
 import type { Meta, StoryObj } from '@storybook/react';
-import Avatar from '@mui/material/Avatar';
+import { Avatar } from '@availity/mui-avatar';
 import { FolderIcon } from '@availity/mui-icon';
+import { Typography } from '@availity/mui-typography';
 import { List, ListItem, ListItemAvatar, ListItemText, ListSubheader, ListSubheaderProps } from '..';
 
 const meta: Meta<typeof ListSubheader> = {
@@ -14,41 +15,44 @@ const meta: Meta<typeof ListSubheader> = {
 export default meta;
 
 export const _ListSubheader: StoryObj<typeof ListSubheader> = {
-  render: (args: ListSubheaderProps) => <ListSubheader {...args} />,
+  render: (args: ListSubheaderProps) => <List><ListSubheader {...args}/></List>,
   args: {
-    children: 'This text is a child of List',
+    children: 'This text is a child of ListSubheader',
   },
 };
 
 export const _FolderList: StoryObj<typeof List> = {
   render: () => (
-    <List sx={{ width: '100%', bgcolor: 'background.paper' }}>
-      <ListSubheader>Subheader 1</ListSubheader>
-      <ListItem>
-        <ListItemAvatar>
-          <Avatar>
-            <FolderIcon />
-          </Avatar>
-        </ListItemAvatar>
-        <ListItemText primary="Photos" secondary="Jan 9, 2014" />
-      </ListItem>
-      <ListItem>
-        <ListItemAvatar>
-          <Avatar>
-            <FolderIcon />
-          </Avatar>
-        </ListItemAvatar>
-        <ListItemText primary="Work" secondary="Jan 7, 2014" />
-      </ListItem>
-      <ListSubheader>Subheader 2</ListSubheader>
-      <ListItem>
-        <ListItemAvatar>
-          <Avatar>
-            <FolderIcon />
-          </Avatar>
-        </ListItemAvatar>
-        <ListItemText primary="Vacation" secondary="July 20, 2014" />
-      </ListItem>
-    </List>
+    <>
+      <Typography variant="h6" component="div">List Header</Typography>
+      <List sx={{ width: '100%', bgcolor: 'background.paper' }}>
+        <ListSubheader>Subheader 1</ListSubheader>
+        <ListItem>
+          <ListItemAvatar>
+            <Avatar size="m">
+              <FolderIcon />
+            </Avatar>
+          </ListItemAvatar>
+          <ListItemText primary="Photos" secondary="Jan 9, 2014" />
+        </ListItem>
+        <ListItem>
+          <ListItemAvatar>
+            <Avatar size="m">
+              <FolderIcon />
+            </Avatar>
+          </ListItemAvatar>
+          <ListItemText primary="Work" secondary="Jan 7, 2014" />
+        </ListItem>
+        <ListSubheader>Subheader 2</ListSubheader>
+        <ListItem>
+          <ListItemAvatar>
+            <Avatar size="m">
+              <FolderIcon />
+            </Avatar>
+          </ListItemAvatar>
+          <ListItemText primary="Vacation" secondary="July 20, 2014" />
+        </ListItem>
+      </List>
+    </>
   )
 };