+++ packages/list/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.2](https://github.com/Availity/element/compare/@availity/mui-list@0.1.1...@availity/mui-list@0.1.2) (2024-07-11)
+
+
+### Bug Fixes
+
+* **mui-list:** ux listitemstatuscard tweaks ([0cb89bf](https://github.com/Availity/element/commit/0cb89bf6e417ff4487b79a3bdc90444b25e2e857))
+* **theme:** ux list tweaks ([a0f303e](https://github.com/Availity/element/commit/a0f303e46e32b08f92c9f933e3dde1bf62b836bc))
+
+## [0.1.1](https://github.com/Availity/element/compare/@availity/mui-list@0.1.0...@availity/mui-list@0.1.1) (2024-07-01)
+
+### Dependency Updates
+
+* `mui-checkbox` updated to version `0.1.0`
 ## 0.1.0 (2024-06-25)
 
 ### Dependency Updates